move patterns to vector-sink
